requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
streamlit>=1.20.0
sqlalchemy>=1.4.0
pymysql>=1.0.0
openai>=1.0.0
sentence-transformers>=2.0.0
numpy>=1.20.0
pandas>=1.3.0
python-dotenv>=0.19.0
pydantic>=1.8.0
httpx>=0.24.0
tiktoken>=0.3.0
python-multipart>=0.0.5
tidb-vector>=0.0.9
strands-agents
strands-agents-builder
strands-agents-tools
mcp>=1.0.0
